+++ handlers.go
@@ -230,7 +230,7 @@ func seasonsHandler(db *sql.DB) http.HandlerFunc {
 				log.Println("season snapshot missing; using live economy snapshot")
 				liveCoinsValue, liveStarsValue, _ := economy.Snapshot()
 				snapshotEnded = now
-				snapshotCoins = liveCoinsValue
+				snapshotCoins = int64(liveCoinsValue)
 				snapshotStars = int64(liveStarsValue)
 			} else if err != nil {
 				log.Println("season snapshot query failed:", err)
